refactor(plugin-thread): merge ChannelCompanion into ChannelArticle
Fold the channel chat into ChannelArticle (exported guard + Panel + local
ChannelCompanion component); the chat-companion surface renders ChannelArticle.
