+++ packages/plugins/plugin-thread/src/capabilities/react-surface.tsx
@@ -11,7 +11,7 @@ import { AppSurface } from '@dxos/app-toolkit/ui';
 import { getSpace } from '@dxos/react-client/echo';
 import { Channel, Thread } from '@dxos/types';
 
-import { ChannelCompanion, ChannelArticle, ThreadArticle } from '#containers';
+import { ChannelArticle, ThreadArticle } from '#containers';
 
 export default Capability.makeModule(() =>
   Effect.succeed(
@@ -29,14 +29,7 @@ export default Capability.makeModule(() =>
           AppSurface.literal(AppSurface.Article, 'chat'),
           AppSurface.companion(AppSurface.Article, Channel.Channel),
         ),
-        component: ({ data: { companionTo: channel } }) => {
-          const space = getSpace(channel);
-          if (!space) {
-            return null;
-          }
-
-          return <ChannelCompanion space={space} channel={channel} />;
-        },
+        component: ({ data: { companionTo: channel } }) => <ChannelArticle subject={channel} />,
       }),
       // TODO(burdon): Disambiguate with Channel.
       Surface.create({
