--- packages/plugins/plugin-thread/src/containers/ChannelCompanion/index.ts
@@ -1,8 +1,0 @@
-//
-// Copyright 2026 DXOS.org
-//
-
-import { ChannelCompanion } from './ChannelCompanion';
-
-export { ChannelCompanion };
-export default ChannelCompanion;
