+++ packages/plugins/plugin-thread/src/containers/ChannelCreatePanel/index.ts
@@ -2,7 +2,4 @@
 // Copyright 2026 DXOS.org
 //
 
-import { ChannelCreatePanel } from './ChannelCreatePanel';
-
-export { ChannelCreatePanel };
-export default ChannelCreatePanel;
+export { ChannelCreatePanel as default } from './ChannelCreatePanel';
