+++ packages/plugins/plugin-thread/src/containers/ThreadArticle/index.ts
@@ -2,7 +2,4 @@
 // Copyright 2023 DXOS.org
 //
 
-import { ThreadArticle } from './ThreadArticle';
-
-export { ThreadArticle };
-export default ThreadArticle;
+export { ThreadArticle as default } from './ThreadArticle';
